third_party/blink/web_tests/external/wpt/css/cssom/resolved-border-width-expected.txt

This is a testharness.js-based test.
[FAIL] checkbox
  assert_equals: border-width expected "5px" but got "0px"
[FAIL] radio
  assert_equals: border-width expected "5px" but got "0px"
Harness: the test ran to completion.
